bugfix: fix overflow for rate limit

+++ common/util/rate_limiter.go
@@ -23,10 +23,10 @@ import (
 
 // RateLimiter is used for limiting the rate of transporting.
 type RateLimiter struct {
-	capacity      int32
-	bucket        int32
-	rate          int32
-	ratePerWindow int32
+	capacity      int64
+	bucket        int64
+	rate          int64
+	ratePerWindow int64
 	window        int64
 	last          int64
 
@@ -37,7 +37,7 @@ type RateLimiter struct {
 // rate: how many tokens are generated per second. 0 represents that don't limit the rate.
 // window: generating tokens interval (millisecond, [1,1000]).
 // The production of rate and window should be division by 1000.
-func NewRateLimiter(rate int32, window int64) *RateLimiter {
+func NewRateLimiter(rate int64, window int64) *RateLimiter {
 	rl := new(RateLimiter)
 	rl.capacity = rate
 	rl.bucket = 0
@@ -50,18 +50,18 @@ func NewRateLimiter(rate int32, window int64) *RateLimiter {
 
 // AcquireBlocking acquires tokens. It will be blocking unit the bucket has enough required
 // number of tokens.
-func (rl *RateLimiter) AcquireBlocking(token int32) int32 {
+func (rl *RateLimiter) AcquireBlocking(token int64) int64 {
 	return rl.acquire(token, true)
 }
 
 // AcquireNonBlocking acquires tokens. It will return -1 immediately when there is no enough
 // number of tokens.
-func (rl *RateLimiter) AcquireNonBlocking(token int32) int32 {
+func (rl *RateLimiter) AcquireNonBlocking(token int64) int64 {
 	return rl.acquire(token, false)
 }
 
 // SetRate sets rate of RateLimiter.
-func (rl *RateLimiter) SetRate(rate int32) {
+func (rl *RateLimiter) SetRate(rate int64) {
 	if rl.rate != rate {
 		rl.capacity = rate
 		rl.rate = rate
@@ -69,14 +69,14 @@ func (rl *RateLimiter) SetRate(rate int32) {
 	}
 }
 
-func (rl *RateLimiter) acquire(token int32, blocking bool) int32 {
+func (rl *RateLimiter) acquire(token int64, blocking bool) int64 {
 	if rl.capacity <= 0 || token < 1 {
 		return token
 	}
 	tmpCapacity := Max(rl.capacity, token)
 
-	var process func() int32
-	process = func() int32 {
+	var process func() int64
+	process = func() int64 {
 		now := time.Now().UnixNano()
 
 		newTokens := rl.createTokens(now)
@@ -113,7 +113,7 @@ func (rl *RateLimiter) computeRatePerWindow() {
 	if rl.rate <= 0 {
 		return
 	}
-	ratePerWindow := int32(int64(rl.rate) * int64(rl.window) / 1000)
+	ratePerWindow := int64(rl.rate) * int64(rl.window) / 1000
 	if ratePerWindow > 0 {
 		rl.ratePerWindow = ratePerWindow
 		return
@@ -122,15 +122,15 @@ func (rl *RateLimiter) computeRatePerWindow() {
 	rl.setWindow(int64(rl.ratePerWindow * 1000 / rl.rate))
 }
 
-func (rl *RateLimiter) createTokens(timeNano int64) int32 {
+func (rl *RateLimiter) createTokens(timeNano int64) int64 {
 	diff := timeNano - rl.last
 	if diff < time.Millisecond.Nanoseconds() {
 		return 0
 	}
-	return int32(diff/(rl.window*time.Millisecond.Nanoseconds())) * rl.ratePerWindow
+	return diff / (rl.window * time.Millisecond.Nanoseconds()) * rl.ratePerWindow
 }
 
-func (rl *RateLimiter) blocking(requiredToken int32) {
+func (rl *RateLimiter) blocking(requiredToken int64) {
 	if requiredToken <= 0 {
 		return
 	}
@@ -140,10 +140,10 @@ func (rl *RateLimiter) blocking(requiredToken int32) {
 
 // TransRate trans the rate to multiples of 1000
 // For NewRateLimiter, the production of rate should be division by 1000.
-func TransRate(rate int) int32 {
+func TransRate(rate int) int64 {
 	if rate <= 0 {
 		rate = 10 * 1024 * 1024
 	}
 	rate = (rate/1000 + 1) * 1000
-	return int32(rate)
+	return int64(rate)
 }
